Moving reconst.b_tensor_utils to io.b_tensor
